Solved 2023 Day 13
- Added run module for helper functions for running parts
- Added DoubleEndedIterator impl for InspectDebug iter
- Fixed DoubleEndedIterator impls for RowIter2D and ColIter2D

+++ src/util/rust/arraytools/array2d.rs
@@ -48,7 +48,7 @@ impl<'a, T: 'a> Array2D<'a, T> for Vec<Vec<T>> {
     }
 
     fn width(&self) -> i32 {
-        self.width_at(0).unwrap_or(0) as i32
+        self.width_at(0).unwrap_or(0)
     }
 
     fn width_at(&self, y: i32) -> Option<i32> {
